refactor(export): split launcher export module
Refs #290 task 12.
