--- src/crosshook-native/crates/crosshook-core/src/export/launcher.rs
@@ -1,1497 +1,0 @@
-use std::env;
-use std::error::Error;
-use std::fmt;
-use std::fs;
-use std::io;
-use std::path::{Path, PathBuf};
-
-use crate::launch::runtime_helpers::build_gamescope_args;
-use crate::launch::script_runner::resolve_launch_proton_path;
-use crate::platform::normalize_flatpak_host_path;
-use crate::profile::{GamescopeConfig, TrainerLoadingMode};
-use crate::settings::UmuPreference;
-use serde::{Deserialize, Serialize};
-
-const TRAINER_SUFFIX: &str = " - Trainer";
-
-#[derive(Debug, Clone, PartialEq, Eq, Default, Serialize, Deserialize)]
-pub struct SteamExternalLauncherExportRequest {
-    pub method: String,
-    pub launcher_name: String,
-    pub trainer_path: String,
-    #[serde(default)]
-    pub trainer_loading_mode: TrainerLoadingMode,
-    pub launcher_icon_path: String,
-    pub prefix_path: String,
-    pub proton_path: String,
-    pub steam_app_id: String,
-    pub steam_client_install_path: String,
-    pub target_home_path: String,
-    #[serde(default)]
-    pub profile_name: Option<String>,
-    #[serde(default)]
-    pub runtime_steam_app_id: String,
-    #[serde(default)]
-    pub umu_game_id: String,
-    #[serde(default)]
-    pub umu_preference: UmuPreference,
-    #[serde(default)]
-    pub network_isolation: bool,
-    #[serde(default)]
-    pub gamescope: GamescopeConfig,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Default, Serialize, Deserialize)]
-pub struct SteamExternalLauncherExportResult {
-    pub display_name: String,
-    pub launcher_slug: String,
-    pub script_path: String,
-    pub desktop_entry_path: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum SteamExternalLauncherExportValidationError {
-    TrainerPathRequired,
-    PrefixPathRequired,
-    ProtonPathRequired,
-    SteamAppIdRequired,
-    TargetHomePathRequired,
-    LauncherIconPathNotFound,
-    LauncherIconPathInvalidExtension,
-    UnsupportedMethod(String),
-}
-
-impl SteamExternalLauncherExportValidationError {
-    pub fn message(&self) -> &'static str {
-        match self {
-            Self::TrainerPathRequired => "External launcher export requires a trainer path.",
-            Self::PrefixPathRequired => "External launcher export requires a prefix path.",
-            Self::ProtonPathRequired => "External launcher export requires a Proton path.",
-            Self::SteamAppIdRequired => "External launcher export requires a Steam App ID.",
-            Self::TargetHomePathRequired => "External launcher export requires a host home path.",
-            Self::LauncherIconPathNotFound => "External launcher export icon path does not exist.",
-            Self::LauncherIconPathInvalidExtension => {
-                "External launcher export icon must be a PNG or JPG image."
-            }
-            Self::UnsupportedMethod(_) => {
-                "External launcher export only supports steam_applaunch and proton_run."
-            }
-        }
-    }
-}
-
-impl fmt::Display for SteamExternalLauncherExportValidationError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::UnsupportedMethod(method) => write!(
-                f,
-                "External launcher export only supports steam_applaunch and proton_run, not '{method}'."
-            ),
-            _ => f.write_str(self.message()),
-        }
-    }
-}
-
-impl Error for SteamExternalLauncherExportValidationError {}
-
-#[derive(Debug)]
-pub enum SteamExternalLauncherExportError {
-    InvalidRequest(SteamExternalLauncherExportValidationError),
-    CouldNotResolveHomePath,
-    Io(io::Error),
-}
-
-impl fmt::Display for SteamExternalLauncherExportError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::InvalidRequest(error) => f.write_str(error.message()),
-            Self::CouldNotResolveHomePath => {
-                f.write_str("Could not resolve a host home path for launcher export.")
-            }
-            Self::Io(error) => write!(f, "{error}"),
-        }
-    }
-}
-
-impl Error for SteamExternalLauncherExportError {
-    fn source(&self) -> Option<&(dyn Error + 'static)> {
-        match self {
-            Self::Io(error) => Some(error),
-            _ => None,
-        }
-    }
-}
-
-impl From<io::Error> for SteamExternalLauncherExportError {
-    fn from(value: io::Error) -> Self {
-        Self::Io(value)
-    }
-}
-
-pub fn validate(
-    request: &SteamExternalLauncherExportRequest,
-) -> Result<(), SteamExternalLauncherExportValidationError> {
-    match request.method.trim() {
-        "steam_applaunch" | "proton_run" => {}
-        other => {
-            return Err(
-                SteamExternalLauncherExportValidationError::UnsupportedMethod(other.to_string()),
-            )
-        }
-    }
-
-    if request.trainer_path.trim().is_empty() {
-        return Err(SteamExternalLauncherExportValidationError::TrainerPathRequired);
-    }
-
-    if request.prefix_path.trim().is_empty() {
-        return Err(SteamExternalLauncherExportValidationError::PrefixPathRequired);
-    }
-
-    if request.proton_path.trim().is_empty() {
-        return Err(SteamExternalLauncherExportValidationError::ProtonPathRequired);
-    }
-
-    if request.target_home_path.trim().is_empty()
-        && env::var("HOME").unwrap_or_default().trim().is_empty()
-    {
-        return Err(SteamExternalLauncherExportValidationError::TargetHomePathRequired);
-    }
-
-    if request.method.trim() == "steam_applaunch" && request.steam_app_id.trim().is_empty() {
-        return Err(SteamExternalLauncherExportValidationError::SteamAppIdRequired);
-    }
-
-    if !request.launcher_icon_path.trim().is_empty() {
-        let normalized_icon_path = normalize_host_unix_path(&request.launcher_icon_path);
-        let icon_path = Path::new(&normalized_icon_path);
-
-        if !icon_path.exists() {
-            return Err(SteamExternalLauncherExportValidationError::LauncherIconPathNotFound);
-        }
-
-        let extension = icon_path
-            .extension()
-            .and_then(|value| value.to_str())
-            .unwrap_or_default();
-
-        if !matches!(
-            extension.to_ascii_lowercase().as_str(),
-            "png" | "jpg" | "jpeg"
-        ) {
-            return Err(
-                SteamExternalLauncherExportValidationError::LauncherIconPathInvalidExtension,
-            );
-        }
-    }
-
-    Ok(())
-}
-
-pub fn export_launchers(
-    request: &SteamExternalLauncherExportRequest,
-) -> Result<SteamExternalLauncherExportResult, SteamExternalLauncherExportError> {
-    validate(request).map_err(SteamExternalLauncherExportError::InvalidRequest)?;
-
-    let display_name = resolve_display_name(
-        &request.launcher_name,
-        &request.steam_app_id,
-        &request.trainer_path,
-    );
-    let launcher_slug = sanitize_launcher_slug(&display_name);
-    let target_home_path = resolve_target_home_path(
-        &request.target_home_path,
-        &request.steam_client_install_path,
-    );
-
-    if target_home_path.trim().is_empty() {
-        return Err(SteamExternalLauncherExportError::CouldNotResolveHomePath);
-    }
-
-    let script_path = combine_host_unix_path(
-        &target_home_path,
-        ".local/share/crosshook/launchers",
-        &format!("{launcher_slug}-trainer.sh"),
-    );
-    let desktop_entry_path = combine_host_unix_path(
-        &target_home_path,
-        ".local/share/applications",
-        &format!("crosshook-{launcher_slug}-trainer.desktop"),
-    );
-
-    write_host_text_file(
-        &script_path,
-        &build_trainer_script_content(request, &display_name),
-        0o755,
-    )?;
-    write_host_text_file(
-        &desktop_entry_path,
-        &build_desktop_entry_content(
-            &display_name,
-            &launcher_slug,
-            &script_path,
-            &request.launcher_icon_path,
-        ),
-        0o644,
-    )?;
-
-    Ok(SteamExternalLauncherExportResult {
-        display_name,
-        launcher_slug,
-        script_path,
-        desktop_entry_path,
-    })
-}
-
-pub(crate) fn resolve_display_name(
-    preferred_name: &str,
-    steam_app_id: &str,
-    trainer_path: &str,
-) -> String {
-    if !preferred_name.trim().is_empty() {
-        return strip_trainer_suffix(preferred_name);
-    }
-
-    let trainer_name = Path::new(trainer_path)
-        .file_stem()
-        .and_then(|value| value.to_str())
-        .unwrap_or_default()
-        .trim();
-
-    if !trainer_name.is_empty() {
-        return strip_trainer_suffix(trainer_name);
-    }
-
-    if !steam_app_id.trim().is_empty() {
-        format!("steam-{steam_app_id}-trainer")
-    } else {
-        "crosshook-trainer".to_string()
-    }
-}
-
-pub(crate) fn strip_trainer_suffix(value: &str) -> String {
-    let trimmed = value.trim();
-    trimmed
-        .strip_suffix(TRAINER_SUFFIX)
-        .unwrap_or(trimmed)
-        .trim_end()
-        .to_string()
-}
-
-pub fn sanitize_launcher_slug(value: &str) -> String {
-    if value.trim().is_empty() {
-        return "crosshook-trainer".to_string();
-    }
-
-    let mut slug = String::with_capacity(value.len());
-    let mut last_character_was_separator = false;
-
-    for character in value.trim().chars().flat_map(char::to_lowercase) {
-        if character.is_alphanumeric() {
-            slug.push(character);
-            last_character_was_separator = false;
-            continue;
-        }
-
-        if last_character_was_separator {
-            continue;
-        }
-
-        slug.push('-');
-        last_character_was_separator = true;
-    }
-
-    let slug = slug.trim_matches('-').to_string();
-    if slug.is_empty() {
-        "crosshook-trainer".to_string()
-    } else {
-        slug
-    }
-}
-
-pub(crate) fn combine_host_unix_path(
-    root_path: &str,
-    segment_one: &str,
-    segment_two: &str,
-) -> String {
-    let normalized_root_path = normalize_host_unix_path(root_path);
-    let normalized_root_path = normalized_root_path.trim_end_matches('/');
-    if normalized_root_path.is_empty() {
-        return String::new();
-    }
-
-    let mut result = normalized_root_path.to_string();
-    for segment in [segment_one, segment_two] {
-        let normalized_segment = normalize_host_unix_path(segment);
-        let normalized_segment = normalized_segment.trim_matches('/');
-        if normalized_segment.is_empty() {
-            continue;
-        }
-
-        result.push('/');
-        result.push_str(normalized_segment);
-    }
-
-    result
-}
-
-pub(crate) fn build_trainer_script_content(
-    request: &SteamExternalLauncherExportRequest,
-    display_name: &str,
-) -> String {
-    let resolved_proton_path =
-        resolve_launch_proton_path(&request.proton_path, &request.steam_client_install_path);
-    let mut content = String::new();
-    content.push_str("#!/usr/bin/env bash\n");
-    content.push_str("set -euo pipefail\n\n");
-    content.push_str(&format!("# {display_name} - Trainer launcher\n"));
-    content.push_str("# Generated by CrossHook\n");
-    content.push_str("# https://github.com/yandy-r/crosshook\n");
-    content.push_str("# Launch this after the game has reached the in-game menu.\n");
-    content.push_str(&format!(
-        "# Trainer loading mode: {}\n",
-        request.trainer_loading_mode.as_str()
-    ));
-    content.push_str(&format!(
-        "# Network isolation: {}\n",
-        if request.network_isolation {
-            "enabled (unshare --net)"
-        } else {
-            "disabled"
-        }
-    ));
-    content.push_str(&format!(
-        "PREFIX_ROOT={}\n",
-        shell_single_quoted(&request.prefix_path)
-    ));
-    content.push_str(
-        r#"if [[ "${PREFIX_ROOT##*/}" == "pfx" ]]; then
-  export WINEPREFIX="$PREFIX_ROOT"
-  export STEAM_COMPAT_DATA_PATH="$(dirname "$PREFIX_ROOT")"
-elif [[ -d "$PREFIX_ROOT/pfx" ]]; then
-  export WINEPREFIX="$PREFIX_ROOT/pfx"
-  export STEAM_COMPAT_DATA_PATH="$PREFIX_ROOT"
-else
-  export WINEPREFIX="$PREFIX_ROOT"
-  export STEAM_COMPAT_DATA_PATH="$PREFIX_ROOT"
-fi
-"#,
-    );
-    if !request.steam_client_install_path.trim().is_empty() {
-        content.push_str(&format!(
-            "export STEAM_COMPAT_CLIENT_INSTALL_PATH={}\n",
-            shell_single_quoted(&request.steam_client_install_path)
-        ));
-    }
-    content.push_str(&format!(
-        "PROTON={}\n",
-        shell_single_quoted(&resolved_proton_path)
-    ));
-    content.push_str(&format!(
-        "TRAINER_HOST_PATH={}\n",
-        shell_single_quoted(&request.trainer_path)
-    ));
-    if request.gamescope.enabled {
-        content.push_str(&build_gamescope_script_block(&request.gamescope));
-    }
-    match request.trainer_loading_mode {
-        TrainerLoadingMode::SourceDirectory => {
-            content.push_str(
-                r#"
-
-trainer_host_path="$(realpath "$TRAINER_HOST_PATH")"
-trainer_host_dir="$(dirname "$trainer_host_path")"
-cd "$trainer_host_dir"
-
-"#,
-            );
-            content.push_str(&build_exec_line(request, "\"$trainer_host_path\""));
-        }
-        TrainerLoadingMode::CopyToPrefix => {
-            content.push_str(
-                r#"
-
-copy_support_directory_if_present() {
-  local source_dir="$1"
-  local target_dir="$2"
-  local child_name="$3"
-
-  if [[ -d "$source_dir/$child_name" ]]; then
-    mkdir -p "$target_dir"
-    cp -R "$source_dir/$child_name" "$target_dir/"
-  fi
-}
-
-stage_trainer_support_files() {
-  local trainer_source_dir="$1"
-  local staged_target_dir="$2"
-  local trainer_file_name="$3"
-  local trainer_base_name="$4"
-  local sibling_file
-  local sibling_name
-
-  shopt -s nullglob
-
-  for sibling_file in "$trainer_source_dir"/*; do
-    sibling_name="$(basename "$sibling_file")"
-
-    if [[ "$sibling_name" == "$trainer_file_name" ]]; then
-      continue
-    fi
-
-    if [[ -f "$sibling_file" ]]; then
-      case "$sibling_name" in
-        "$trainer_base_name".*.json|\
-        "$trainer_base_name".*.config|\
-        "$trainer_base_name".*.ini|\
-        "$trainer_base_name".*.dll|\
-        "$trainer_base_name".*.bin|\
-        "$trainer_base_name".*.dat|\
-        "$trainer_base_name".*.pak)
-          cp -f "$sibling_file" "$staged_target_dir/"
-          ;;
-        *.dll|*.json|*.config|*.ini|*.pak|*.dat|*.bin)
-          cp -f "$sibling_file" "$staged_target_dir/"
-          ;;
-      esac
-    fi
-  done
-
-  shopt -u nullglob
-
-  for support_dir in assets data lib bin runtimes plugins locales cef resources; do
-    copy_support_directory_if_present "$trainer_source_dir" "$staged_target_dir" "$support_dir"
-  done
-}
-
-trainer_host_path="$(realpath "$TRAINER_HOST_PATH")"
-trainer_file_name="$(basename "$trainer_host_path")"
-trainer_base_name="${trainer_file_name%.*}"
-trainer_source_dir="$(dirname "$trainer_host_path")"
-staged_trainer_root="$WINEPREFIX/drive_c/CrossHook/StagedTrainers"
-staged_trainer_dir="$staged_trainer_root/$trainer_base_name"
-staged_trainer_host_path="$staged_trainer_dir/$trainer_file_name"
-staged_trainer_windows_path="C:\\CrossHook\\StagedTrainers\\$trainer_base_name\\$trainer_file_name"
-
-mkdir -p "$staged_trainer_dir"
-cp -f "$trainer_host_path" "$staged_trainer_host_path"
-stage_trainer_support_files "$trainer_source_dir" "$staged_trainer_dir" "$trainer_file_name" "$trainer_base_name"
-
-"#,
-            );
-            content.push_str(&build_exec_line(
-                request,
-                "\"$staged_trainer_windows_path\"",
-            ));
-        }
-    }
-    content
-}
-
-fn build_gamescope_script_block(config: &GamescopeConfig) -> String {
-    let raw_args = build_gamescope_args(config);
-    let has_fullscreen = raw_args.iter().any(|a| a == "-f");
-
-    let mut quoted_args: Vec<String> = Vec::new();
-    if !has_fullscreen {
-        quoted_args.push(shell_single_quoted("-f"));
-    }
-    for token in &raw_args {
-        quoted_args.push(shell_single_quoted(token));
-    }
-    let args_literal = quoted_args.join(" ");
-
-    let mut block = String::new();
-    block.push_str(&format!(
-        "\n# Gamescope wrapper — launches trainer in its own compositor session.\n\
-         _GAMESCOPE_ARGS=({args_literal})\n"
-    ));
-
-    if config.allow_nested {
-        block.push_str("_GS_PREFIX=(gamescope \"${_GAMESCOPE_ARGS[@]}\" --)\n");
-    } else {
-        block.push_str(
-            r#"if [[ -z "${GAMESCOPE_WAYLAND_DISPLAY:-}" ]]; then
-  _GS_PREFIX=(gamescope "${_GAMESCOPE_ARGS[@]}" --)
-else
-  _GS_PREFIX=()
-fi
-"#,
-        );
-    }
-
-    block
-}
-
-fn build_exec_line(request: &SteamExternalLauncherExportRequest, target_path: &str) -> String {
-    let mut block = String::new();
-
-    if request.network_isolation {
-        // Runtime probe: try `unshare --net true`; fall back with a warning if it fails.
-        block.push_str(
-            r#"_NET_PREFIX=()
-if unshare --net true >/dev/null 2>&1; then
-  _NET_PREFIX=(unshare --net)
-else
-  echo "[CrossHook] WARNING: unshare --net unavailable — launching without network isolation" >&2
-fi
-"#,
-        );
-    }
-
-    // Build the shell-array prefix tokens based on which wrappers are active.
-    // GOTCHA: when both gamescope_enabled and network_isolation are false,
-    // neither _GS_PREFIX nor _NET_PREFIX is declared earlier — emit no tokens.
-    let prefix = match (request.gamescope.enabled, request.network_isolation) {
-        (true, true) => "\"${_GS_PREFIX[@]}\" \"${_NET_PREFIX[@]}\" ",
-        (true, false) => "\"${_GS_PREFIX[@]}\" ",
-        (false, true) => "\"${_NET_PREFIX[@]}\" ",
-        (false, false) => "",
-    };
-    let direct_proton_exec = format!(
-        "unset GAMEID PROTON_VERB PROTONPATH\nexec {prefix}\"$PROTON\" run {target_path}\n"
-    );
-
-    match request.umu_preference {
-        UmuPreference::Proton => {
-            block.push_str(&direct_proton_exec);
-        }
-        UmuPreference::Auto | UmuPreference::Umu => {
-            // Probe umu-run once; Auto fallback is silent.
-            block.push_str(
-                r#"_UMU_AVAILABLE=0
-if command -v umu-run >/dev/null 2>&1; then
-  _UMU_AVAILABLE=1
-fi
-"#,
-            );
-            let umu_exports = build_umu_env_exports(request);
-            let fallback_warning = if request.umu_preference == UmuPreference::Umu {
-                "  echo \"[CrossHook] WARNING: umu preference requested but umu-run is unavailable — falling back to direct Proton\" >&2\n"
-            } else {
-                ""
-            };
-            block.push_str(&format!(
-                "if [ \"$_UMU_AVAILABLE\" = \"1\" ]; then\n{umu_exports}  exec {prefix}umu-run {target_path}\nelse\n{fallback_warning}  {direct_proton_exec}fi\n"
-            ));
-        }
-    }
-
-    block
-}
-
-fn build_umu_env_exports(request: &SteamExternalLauncherExportRequest) -> String {
-    let resolved_proton_path =
-        resolve_launch_proton_path(&request.proton_path, &request.steam_client_install_path);
-    format!(
-        "  export GAMEID={}\n  export PROTON_VERB='runinprefix'\n  export PROTONPATH={}\n",
-        shell_single_quoted(&resolve_umu_game_id_for_env(request)),
-        shell_single_quoted(&proton_path_dirname(&resolved_proton_path))
-    )
-}
-
-fn resolve_steam_app_id_for_umu(request: &SteamExternalLauncherExportRequest) -> &str {
-    let umu_override = request.umu_game_id.trim();
-    if !umu_override.is_empty() {
-        return umu_override;
-    }
-
-    let steam_id = request.steam_app_id.trim();
-    if !steam_id.is_empty() {
-        return steam_id;
-    }
-
-    let runtime_id = request.runtime_steam_app_id.trim();
-    if !runtime_id.is_empty() {
-        return runtime_id;
-    }
-
-    ""
-}
-
-fn resolve_umu_game_id_for_env(request: &SteamExternalLauncherExportRequest) -> String {
-    let trimmed = resolve_steam_app_id_for_umu(request).trim();
-    if trimmed.is_empty() {
-        "umu-0".to_string()
-    } else {
-        trimmed.to_string()
-    }
-}
-
-fn proton_path_dirname(proton_path: &str) -> String {
-    Path::new(proton_path.trim())
-        .parent()
-        .map(|parent| parent.to_string_lossy().into_owned())
-        .unwrap_or_default()
-}
-
-pub(crate) fn build_desktop_entry_content(
-    display_name: &str,
-    slug: &str,
-    script_path: &str,
-    launcher_icon_path: &str,
-) -> String {
-    let mut content = String::new();
-    content.push_str("[Desktop Entry]\n");
-    content.push_str("Type=Application\n");
-    content.push_str("Version=1.0\n");
-    content.push_str(&format!("Name={display_name} - Trainer\n"));
-    content.push_str(&format!(
-        "Comment=Trainer launcher for {display_name}. Generated by CrossHook: https://github.com/yandy-r/crosshook\n"
-    ));
-    content.push_str(&format!(
-        "Exec=/bin/bash {}\n",
-        escape_desktop_exec_argument(script_path)
-    ));
-    content.push_str("Terminal=false\n");
-    content.push_str("Categories=Game;\n");
-    content.push_str(&format!(
-        "Icon={}\n",
-        resolve_desktop_icon_value(launcher_icon_path)
-    ));
-    content.push_str("StartupNotify=false\n");
-    content.push_str(&format!("X-CrossHook-Profile={display_name}\n"));
-    content.push_str(&format!("X-CrossHook-Slug={slug}\n"));
-    content
-}
-
-pub(crate) fn write_host_text_file(
-    host_path: &str,
-    content: &str,
-    mode: u32,
-) -> Result<(), io::Error> {
-    let writable_path = PathBuf::from(host_path);
-    let directory_path = writable_path.parent().ok_or_else(|| {
-        io::Error::new(
-            io::ErrorKind::InvalidInput,
-            format!("could not resolve a parent directory for '{host_path}'"),
-        )
-    })?;
-
-    fs::create_dir_all(directory_path)?;
-    fs::write(&writable_path, content.replace("\r\n", "\n"))?;
-
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::PermissionsExt;
-
-        let mut permissions = fs::metadata(&writable_path)?.permissions();
-        permissions.set_mode(mode);
-        fs::set_permissions(&writable_path, permissions)?;
-    }
-
-    Ok(())
-}
-
-pub fn resolve_target_home_path(
-    preferred_home_path: &str,
-    steam_client_install_path: &str,
-) -> String {
-    let normalized_preferred_home_path = normalize_host_unix_path(preferred_home_path);
-    if looks_like_usable_host_unix_path(&normalized_preferred_home_path) {
-        return normalized_preferred_home_path;
-    }
-
-    let normalized_steam_client_install_path = normalize_host_unix_path(steam_client_install_path);
-    if let Some(derived_home_path) =
-        try_resolve_home_from_steam_client_install_path(&normalized_steam_client_install_path)
-    {
-        return derived_home_path;
-    }
-
-    if let Ok(home_path) = env::var("HOME") {
-        let normalized_home_path = normalize_host_unix_path(&home_path);
-        if looks_like_usable_host_unix_path(&normalized_home_path) {
-            return normalized_home_path;
-        }
-    }
-
-    normalized_preferred_home_path
-}
-
-fn try_resolve_home_from_steam_client_install_path(
-    steam_client_install_path: &str,
-) -> Option<String> {
-    const LOCAL_SHARE_STEAM_SUFFIX: &str = "/.local/share/Steam";
-    const DOT_STEAM_ROOT_SUFFIX: &str = "/.steam/root";
-
-    if steam_client_install_path.trim().is_empty() {
-        return None;
-    }
-
-    if let Some(home_path) = steam_client_install_path.strip_suffix(LOCAL_SHARE_STEAM_SUFFIX) {
-        let home_path = home_path.trim();
-        if !home_path.is_empty() {
-            return Some(home_path.to_string());
-        }
-    }
-
-    if let Some(home_path) = steam_client_install_path.strip_suffix(DOT_STEAM_ROOT_SUFFIX) {
-        let home_path = home_path.trim();
-        if !home_path.is_empty() {
-            return Some(home_path.to_string());
-        }
-    }
-
-    None
-}
-
-fn resolve_desktop_icon_value(launcher_icon_path: &str) -> String {
-    let normalized_launcher_icon_path = normalize_host_unix_path(launcher_icon_path);
-    if normalized_launcher_icon_path.trim().is_empty() {
-        "applications-games".to_string()
-    } else {
-        normalized_launcher_icon_path
-    }
-}
-
-fn normalize_host_unix_path(path: &str) -> String {
-    let normalized = path.trim().replace('\\', "/");
-    normalize_flatpak_host_path(&normalized)
-}
-
-fn looks_like_usable_host_unix_path(path: &str) -> bool {
-    !path.trim().is_empty() && path.starts_with('/') && !path.contains("/compatdata/")
-}
-
-fn shell_single_quoted(value: &str) -> String {
-    format!("'{}'", value.replace('\'', "'\"'\"'"))
-}
-
-fn escape_desktop_exec_argument(value: &str) -> String {
-    value
-        .replace('\\', "\\\\")
-        .replace('%', "%%")
-        .replace(' ', "\\ ")
-        .replace('"', "\\\"")
-}
-
-/// Generates the trainer launcher script content with placement comment headers,
-/// suitable for clipboard copy. Does NOT write to disk.
-pub fn preview_trainer_script_content(
-    request: &SteamExternalLauncherExportRequest,
-) -> Result<String, SteamExternalLauncherExportError> {
-    validate(request).map_err(SteamExternalLauncherExportError::InvalidRequest)?;
-
-    let display_name = resolve_display_name(
-        &request.launcher_name,
-        &request.steam_app_id,
-        &request.trainer_path,
-    );
-    let launcher_slug = sanitize_launcher_slug(&display_name);
-    let target_home_path = resolve_target_home_path(
-        &request.target_home_path,
-        &request.steam_client_install_path,
-    );
-
-    if target_home_path.trim().is_empty() {
-        return Err(SteamExternalLauncherExportError::CouldNotResolveHomePath);
-    }
-
-    let script_path = combine_host_unix_path(
-        &target_home_path,
-        ".local/share/crosshook/launchers",
-        &format!("{launcher_slug}-trainer.sh"),
-    );
-
-    let body = build_trainer_script_content(request, &display_name);
-    Ok(format!(
-        "# Save this file to: {script_path}\n\
-         # Make executable: chmod +x {script_path}\n\
-         \n\
-         {body}"
-    ))
-}
-
-/// Generates the desktop entry content with placement comment headers,
-/// suitable for clipboard copy. Does NOT write to disk.
-pub fn preview_desktop_entry_content(
-    request: &SteamExternalLauncherExportRequest,
-) -> Result<String, SteamExternalLauncherExportError> {
-    validate(request).map_err(SteamExternalLauncherExportError::InvalidRequest)?;
-
-    let display_name = resolve_display_name(
-        &request.launcher_name,
-        &request.steam_app_id,
-        &request.trainer_path,
-    );
-    let launcher_slug = sanitize_launcher_slug(&display_name);
-    let target_home_path = resolve_target_home_path(
-        &request.target_home_path,
-        &request.steam_client_install_path,
-    );
-
-    if target_home_path.trim().is_empty() {
-        return Err(SteamExternalLauncherExportError::CouldNotResolveHomePath);
-    }
-
-    let script_path = combine_host_unix_path(
-        &target_home_path,
-        ".local/share/crosshook/launchers",
-        &format!("{launcher_slug}-trainer.sh"),
-    );
-    let desktop_entry_path = combine_host_unix_path(
-        &target_home_path,
-        ".local/share/applications",
-        &format!("crosshook-{launcher_slug}-trainer.desktop"),
-    );
-
-    let body = build_desktop_entry_content(
-        &display_name,
-        &launcher_slug,
-        &script_path,
-        &request.launcher_icon_path,
-    );
-    Ok(format!(
-        "# Save this file to: {desktop_entry_path}\n\
-         # Permissions: 644 (not executable)\n\
-         \n\
-         {body}"
-    ))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::fs;
-    use tempfile::tempdir;
-
-    #[test]
-    fn slug_generation_collapses_non_alphanumeric_runs() {
-        assert_eq!(
-            sanitize_launcher_slug("  CrossHook: Trainer 2026!!  "),
-            "crosshook-trainer-2026"
-        );
-        assert_eq!(sanitize_launcher_slug(""), "crosshook-trainer");
-        assert_eq!(sanitize_launcher_slug("---"), "crosshook-trainer");
-    }
-
-    #[test]
-    fn shell_single_quote_escaping_matches_posix_pattern() {
-        assert_eq!(shell_single_quoted("abc"), "'abc'");
-        assert_eq!(shell_single_quoted("a'b"), "'a'\"'\"'b'");
-    }
-
-    #[test]
-    fn desktop_exec_escaping_follows_freedesktop_spec() {
-        assert_eq!(
-            escape_desktop_exec_argument("/tmp/Cross Hook/runner\".sh"),
-            "/tmp/Cross\\ Hook/runner\\\".sh"
-        );
-    }
-
-    #[test]
-    fn desktop_exec_escaping_doubles_percent_signs() {
-        assert_eq!(
-            escape_desktop_exec_argument("/opt/trainers/%u/trainer.exe"),
-            "/opt/trainers/%%u/trainer.exe"
-        );
-    }
-
-    #[test]
-    fn normalize_host_unix_path_rewrites_flatpak_host_mount_paths() {
-        assert_eq!(
-            normalize_host_unix_path("/run/host/home/alice/.local/share/icons/cover.png"),
-            "/home/alice/.local/share/icons/cover.png"
-        );
-    }
-
-    #[test]
-    fn normalize_host_unix_path_leaves_regular_host_paths_unchanged() {
-        let path = "/home/alice/.local/share/icons/cover.png";
-        assert_eq!(normalize_host_unix_path(path), path);
-    }
-
-    #[test]
-    fn export_writes_expected_paths_and_content() {
-        let temp_home = tempdir().expect("temp home");
-        let icon_path = temp_home.path().join("launcher icon.png");
-        fs::write(&icon_path, b"icon").expect("icon");
-
-        let request = SteamExternalLauncherExportRequest {
-            method: "steam_applaunch".to_string(),
-            launcher_name: "Elden Ring Deluxe".to_string(),
-            trainer_path: "/opt/Trainers/Trainer's Edition.exe".to_string(),
-            trainer_loading_mode: TrainerLoadingMode::CopyToPrefix,
-            launcher_icon_path: icon_path.to_string_lossy().into_owned(),
-            prefix_path: "/tmp/compatdata/1245620".to_string(),
-            proton_path: "/opt/Proton/proton".to_string(),
-            steam_app_id: "1245620".to_string(),
-            steam_client_install_path: temp_home
-                .path()
-                .join(".local/share/Steam")
-                .to_string_lossy()
-                .into_owned(),
-            target_home_path: "/tmp/not-a-home/compatdata/steam".to_string(),
-            profile_name: None,
-            ..Default::default()
-        };
-
-        let result = export_launchers(&request).expect("export");
-
-        assert_eq!(result.display_name, "Elden Ring Deluxe");
-        assert_eq!(result.launcher_slug, "elden-ring-deluxe");
-        assert_eq!(
-            result.script_path,
-            temp_home
-                .path()
-                .join(".local/share/crosshook/launchers/elden-ring-deluxe-trainer.sh")
-                .to_string_lossy()
-                .into_owned()
-        );
-        assert_eq!(
-            result.desktop_entry_path,
-            temp_home
-                .path()
-                .join(".local/share/applications/crosshook-elden-ring-deluxe-trainer.desktop")
-                .to_string_lossy()
-                .into_owned()
-        );
-
-        let script_content = fs::read_to_string(&result.script_path).expect("script");
-        assert!(script_content.contains("PREFIX_ROOT='/tmp/compatdata/1245620'"));
-        assert!(script_content.contains("elif [[ -d \"$PREFIX_ROOT/pfx\" ]]; then"));
-        assert!(script_content.contains("export WINEPREFIX=\"$PREFIX_ROOT/pfx\""));
-        assert!(script_content.contains("export STEAM_COMPAT_DATA_PATH=\"$PREFIX_ROOT\""));
-        assert!(script_content.contains("export STEAM_COMPAT_CLIENT_INSTALL_PATH='"));
-        assert!(script_content.contains("PROTON='/opt/Proton/proton'"));
-        assert!(script_content
-            .contains("TRAINER_HOST_PATH='/opt/Trainers/Trainer'\"'\"'s Edition.exe'"));
-        assert!(script_content
-            .contains("staged_trainer_root=\"$WINEPREFIX/drive_c/CrossHook/StagedTrainers\""));
-        assert!(script_content.contains("staged_trainer_windows_path=\"C:\\\\CrossHook\\\\StagedTrainers\\\\$trainer_base_name\\\\$trainer_file_name\""));
-        assert!(script_content.contains(r#"exec "$PROTON" run "$staged_trainer_windows_path""#));
-        assert!(script_content.contains(r#"exec umu-run "$staged_trainer_windows_path""#));
-
-        let desktop_content = fs::read_to_string(&result.desktop_entry_path).expect("desktop");
-        assert!(desktop_content.contains("Name=Elden Ring Deluxe - Trainer"));
-        assert!(desktop_content.contains("Exec=/bin/bash "));
-        assert!(desktop_content.contains("Icon="));
-        assert!(desktop_content.contains(&icon_path.to_string_lossy().replace('\\', "\\\\")));
-        assert!(desktop_content.contains("X-CrossHook-Profile=Elden Ring Deluxe\n"));
-        assert!(desktop_content.contains("X-CrossHook-Slug=elden-ring-deluxe\n"));
-
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-
-            let script_mode = fs::metadata(&result.script_path)
-                .expect("script metadata")
-                .permissions()
-                .mode()
-                & 0o777;
-            assert_eq!(script_mode, 0o755, "scripts should be executable");
-
-            let desktop_mode = fs::metadata(&result.desktop_entry_path)
-                .expect("desktop metadata")
-                .permissions()
-                .mode()
-                & 0o777;
-            assert_eq!(
-                desktop_mode, 0o644,
-                ".desktop files should not be executable"
-            );
-        }
-    }
-
-    #[test]
-    fn export_does_not_duplicate_trainer_suffix_when_name_already_has_it() {
-        let temp_home = tempdir().expect("temp home");
-        let request = SteamExternalLauncherExportRequest {
-            method: "steam_applaunch".to_string(),
-            launcher_name: "Elden Ring - Trainer".to_string(),
-            trainer_path: "/opt/trainers/Elden Ring.exe".to_string(),
-            trainer_loading_mode: TrainerLoadingMode::CopyToPrefix,
-            launcher_icon_path: String::new(),
-            prefix_path: "/tmp/compatdata/1245620".to_string(),
-            proton_path: "/opt/Proton/proton".to_string(),
-            steam_app_id: "1245620".to_string(),
-            steam_client_install_path: temp_home
-                .path()
-                .join(".local/share/Steam")
-                .to_string_lossy()
-                .into_owned(),
-            target_home_path: temp_home.path().to_string_lossy().into_owned(),
-            profile_name: None,
-            ..Default::default()
-        };
-
-        let result = export_launchers(&request).expect("export");
-        assert_eq!(result.display_name, "Elden Ring");
-        assert_eq!(result.launcher_slug, "elden-ring");
-
-        let desktop_content = fs::read_to_string(&result.desktop_entry_path).expect("desktop");
-        assert!(desktop_content.contains("Name=Elden Ring - Trainer\n"));
-        assert!(!desktop_content.contains("Name=Elden Ring - Trainer - Trainer\n"));
-    }
-
-    #[test]
-    fn resolve_display_name_strips_trainer_suffix_from_trainer_stem() {
-        assert_eq!(
-            resolve_display_name("", "1245620", "/opt/trainers/Game Name - Trainer.exe"),
-            "Game Name"
-        );
-    }
-
-    #[test]
-    fn desktop_icon_falls_back_to_applications_games() {
-        let content = build_desktop_entry_content("Test", "test", "/tmp/launcher.sh", "");
-        assert!(content.contains("Icon=applications-games"));
-    }
-
-    #[test]
-    fn desktop_entry_contains_crosshook_metadata_lines() {
-        let content = build_desktop_entry_content(
-            "Elden Ring Deluxe",
-            "elden-ring-deluxe",
-            "/tmp/launcher.sh",
-            "",
-        );
-        assert!(content.contains("X-CrossHook-Profile=Elden Ring Deluxe\n"));
-        assert!(content.contains("X-CrossHook-Slug=elden-ring-deluxe\n"));
-    }
-
-    #[test]
-    fn resolves_home_from_steam_client_suffix() {
-        assert_eq!(
-            resolve_target_home_path(
-                "/tmp/wrong/compatdata/steam",
-                "/home/user/.local/share/Steam"
-            ),
-            "/home/user"
-        );
-    }
-
-    #[test]
-    fn proton_run_export_writes_generic_prefix_bootstrap() {
-        let request = SteamExternalLauncherExportRequest {
-            method: "proton_run".to_string(),
-            launcher_name: "Witcher 3".to_string(),
-            trainer_path: "/opt/trainers/Aurora.exe".to_string(),
-            trainer_loading_mode: TrainerLoadingMode::SourceDirectory,
-            launcher_icon_path: String::new(),
-            prefix_path: "/games/prefixes/the-witcher-3".to_string(),
-            proton_path: "/opt/proton/proton".to_string(),
-            steam_app_id: String::new(),
-            steam_client_install_path: String::new(),
-            target_home_path: "/home/user".to_string(),
-            profile_name: None,
-            ..Default::default()
-        };
-
-        let script_content = build_trainer_script_content(&request, "Witcher 3");
-        assert!(script_content.contains("PREFIX_ROOT='/games/prefixes/the-witcher-3'"));
-        assert!(script_content.contains("elif [[ -d \"$PREFIX_ROOT/pfx\" ]]; then"));
-        assert!(script_content.contains("trainer_host_path=\"$(realpath \"$TRAINER_HOST_PATH\")\""));
-        assert!(script_content.contains(r#"exec "$PROTON" run "$trainer_host_path""#));
-        assert!(script_content.contains(r#"exec umu-run "$trainer_host_path""#));
-        assert!(!script_content
-            .contains("staged_trainer_root=\"$WINEPREFIX/drive_c/CrossHook/StagedTrainers\""));
-        assert!(!script_content.contains("export STEAM_COMPAT_CLIENT_INSTALL_PATH="));
-    }
-
-    #[test]
-    fn steam_export_trainer_script_includes_gamescope_when_request_carries_effective_game_config() {
-        let request = SteamExternalLauncherExportRequest {
-            method: "steam_applaunch".to_string(),
-            launcher_name: "Hitman".to_string(),
-            trainer_path: "/games/trainers/t.exe".to_string(),
-            trainer_loading_mode: TrainerLoadingMode::SourceDirectory,
-            launcher_icon_path: String::new(),
-            prefix_path: "/steam/compatdata/123".to_string(),
-            proton_path: "/opt/proton".to_string(),
-            steam_app_id: "863550".to_string(),
-            steam_client_install_path: "/home/u/.local/share/Steam".to_string(),
-            target_home_path: "/home/user".to_string(),
-            profile_name: None,
-            runtime_steam_app_id: String::new(),
-            umu_game_id: String::new(),
-            umu_preference: UmuPreference::Auto,
-            network_isolation: false,
-            gamescope: GamescopeConfig {
-                enabled: true,
-                fullscreen: true,
-                ..Default::default()
-            },
-        };
-
-        let script_content = build_trainer_script_content(&request, "Hitman");
-        assert!(
-            script_content.contains("Gamescope wrapper"),
-            "expected gamescope block in exported script"
-        );
-        assert!(script_content.contains("_GAMESCOPE_ARGS"));
-    }
-
-    #[test]
-    fn preview_script_contains_placement_header_before_shebang() {
-        let request = SteamExternalLauncherExportRequest {
-            method: "proton_run".to_string(),
-            launcher_name: "Witcher 3".to_string(),
-            trainer_path: "/opt/trainers/Aurora.exe".to_string(),
-            trainer_loading_mode: TrainerLoadingMode::SourceDirectory,
-            launcher_icon_path: String::new(),
-            prefix_path: "/games/prefixes/the-witcher-3".to_string(),
-            proton_path: "/opt/proton/proton".to_string(),
-            steam_app_id: String::new(),
-            steam_client_install_path: String::new(),
-            target_home_path: "/home/user".to_string(),
-            profile_name: None,
-            ..Default::default()
-        };
-
-        let content = preview_trainer_script_content(&request).expect("preview script");
-        assert!(content.starts_with("# Save this file to: /home/user/.local/share/crosshook/launchers/witcher-3-trainer.sh\n"));
-        assert!(content.contains("# Make executable: chmod +x"));
-        assert!(content.contains("#!/usr/bin/env bash"));
-    }
-
-    #[test]
-    fn preview_desktop_contains_placement_header_before_entry() {
-        let request = SteamExternalLauncherExportRequest {
-            method: "proton_run".to_string(),
-            launcher_name: "Witcher 3".to_string(),
-            trainer_path: "/opt/trainers/Aurora.exe".to_string(),
-            trainer_loading_mode: TrainerLoadingMode::SourceDirectory,
-            launcher_icon_path: String::new(),
-            prefix_path: "/games/prefixes/the-witcher-3".to_string(),
-            proton_path: "/opt/proton/proton".to_string(),
-            steam_app_id: String::new(),
-            steam_client_install_path: String::new(),
-            target_home_path: "/home/user".to_string(),
-            profile_name: None,
-            ..Default::default()
-        };
-
-        let content = preview_desktop_entry_content(&request).expect("preview desktop");
-        assert!(content.starts_with("# Save this file to: /home/user/.local/share/applications/crosshook-witcher-3-trainer.desktop\n"));
-        assert!(content.contains("# Permissions: 644"));
-        assert!(content.contains("[Desktop Entry]"));
-    }
-
-    #[test]
-    fn preview_script_fails_when_trainer_path_empty() {
-        let request = SteamExternalLauncherExportRequest {
-            method: "proton_run".to_string(),
-            trainer_path: String::new(),
-            prefix_path: "/tmp/prefix".to_string(),
-            proton_path: "/tmp/proton".to_string(),
-            target_home_path: "/home/user".to_string(),
-            ..Default::default()
-        };
-
-        let result = preview_trainer_script_content(&request);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn preview_desktop_fails_when_trainer_path_empty() {
-        let request = SteamExternalLauncherExportRequest {
-            method: "proton_run".to_string(),
-            trainer_path: String::new(),
-            prefix_path: "/tmp/prefix".to_string(),
-            proton_path: "/tmp/proton".to_string(),
-            target_home_path: "/home/user".to_string(),
-            ..Default::default()
-        };
-
-        let result = preview_desktop_entry_content(&request);
-        assert!(result.is_err());
-    }
-
-    fn make_gamescope_request(
-        gamescope: GamescopeConfig,
-        loading_mode: TrainerLoadingMode,
-    ) -> SteamExternalLauncherExportRequest {
-        SteamExternalLauncherExportRequest {
-            method: "proton_run".to_string(),
-            trainer_path: "/opt/trainers/MyGame.exe".to_string(),
-            trainer_loading_mode: loading_mode,
-            prefix_path: "/tmp/compatdata/12345".to_string(),
-            proton_path: "/opt/proton/proton".to_string(),
-            target_home_path: "/home/user".to_string(),
-            gamescope,
-            ..Default::default()
-        }
-    }
-
-    #[test]
-    fn gamescope_disabled_script_unchanged() {
-        let request = make_gamescope_request(
-            GamescopeConfig::default(),
-            TrainerLoadingMode::SourceDirectory,
-        );
-        let content = build_trainer_script_content(&request, "Test Game");
-        assert!(!content.contains("_GAMESCOPE_ARGS"));
-        assert!(!content.contains("_GS_PREFIX"));
-        assert!(!content.contains("gamescope"));
-        assert!(content.contains(r#"exec "$PROTON" run "$trainer_host_path""#));
-        assert!(content.contains(r#"exec umu-run "$trainer_host_path""#));
-    }
-
-    #[test]
-    fn gamescope_enabled_source_directory_wraps_exec() {
-        let request = make_gamescope_request(
-            GamescopeConfig {
-                enabled: true,
-                internal_width: Some(800),
-                internal_height: Some(400),
-                fullscreen: true,
-                ..Default::default()
-            },
-            TrainerLoadingMode::SourceDirectory,
-        );
-        let content = build_trainer_script_content(&request, "Test Game");
-        assert!(content.contains("_GAMESCOPE_ARGS=("));
-        assert!(content.contains("'-f'"));
-        assert!(content.contains("'-w' '800'"));
-        assert!(content.contains("'-h' '400'"));
-        assert!(content.contains("GAMESCOPE_WAYLAND_DISPLAY"));
-        assert!(content.contains(r#"exec "${_GS_PREFIX[@]}" "$PROTON" run "$trainer_host_path""#));
-        assert!(content.contains(r#"exec "${_GS_PREFIX[@]}" umu-run "$trainer_host_path""#));
-    }
-
-    #[test]
-    fn gamescope_enabled_copy_to_prefix_wraps_exec() {
-        let request = make_gamescope_request(
-            GamescopeConfig {
-                enabled: true,
-                fullscreen: true,
-                ..Default::default()
-            },
-            TrainerLoadingMode::CopyToPrefix,
-        );
-        let content = build_trainer_script_content(&request, "Test Game");
-        assert!(content.contains("_GAMESCOPE_ARGS=("));
-        assert!(content
-            .contains(r#"exec "${_GS_PREFIX[@]}" "$PROTON" run "$staged_trainer_windows_path""#));
-        assert!(
-            content.contains(r#"exec "${_GS_PREFIX[@]}" umu-run "$staged_trainer_windows_path""#)
-        );
-    }
-
-    #[test]
-    fn gamescope_allow_nested_skips_guard() {
-        let request = make_gamescope_request(
-            GamescopeConfig {
-                enabled: true,
-                allow_nested: true,
-                fullscreen: true,
-                ..Default::default()
-            },
-            TrainerLoadingMode::SourceDirectory,
-        );
-        let content = build_trainer_script_content(&request, "Test Game");
-        assert!(!content.contains("GAMESCOPE_WAYLAND_DISPLAY"));
-        assert!(content.contains("_GS_PREFIX=("));
-    }
-
-    #[test]
-    fn gamescope_fullscreen_always_present() {
-        let request = make_gamescope_request(
-            GamescopeConfig {
-                enabled: true,
-                fullscreen: false,
-                internal_width: Some(1280),
-                internal_height: Some(720),
-                ..Default::default()
-            },
-            TrainerLoadingMode::SourceDirectory,
-        );
-        let content = build_trainer_script_content(&request, "Test Game");
-        assert!(
-            content.contains("'-f'"),
-            "fullscreen flag must always be present for trainer gamescope"
-        );
-    }
-
-    #[test]
-    fn gamescope_script_does_not_override_display() {
-        let request = make_gamescope_request(
-            GamescopeConfig {
-                enabled: true,
-                fullscreen: true,
-                ..Default::default()
-            },
-            TrainerLoadingMode::SourceDirectory,
-        );
-        let content = build_trainer_script_content(&request, "Test Game");
-        assert!(
-            !content.contains("pgrep -x wineserver"),
-            "script should not detect wineserver DISPLAY"
-        );
-        assert!(
-            !content.contains("DISPLAY="),
-            "script should not override DISPLAY"
-        );
-        assert!(content.contains("gamescope"));
-    }
-
-    #[test]
-    fn network_isolation_enabled_generates_runtime_probe_and_exec() {
-        let mut request = make_gamescope_request(
-            GamescopeConfig::default(),
-            TrainerLoadingMode::SourceDirectory,
-        );
-        request.network_isolation = true;
-        let content = build_trainer_script_content(&request, "Test Game");
-        assert!(
-            content.contains("# Network isolation: enabled"),
-            "script header should indicate network isolation: {content}"
-        );
-        assert!(
-            content.contains("if unshare --net true"),
-            "script should probe unshare availability: {content}"
-        );
-        assert!(
-            content.contains("_NET_PREFIX=(unshare --net)"),
-            "script should set _NET_PREFIX on success: {content}"
-        );
-        assert!(
-            content.contains("WARNING: unshare --net unavailable"),
-            "script should warn on failure: {content}"
-        );
-        assert!(
-            content.contains(r#"exec "${_NET_PREFIX[@]}" "$PROTON" run"#),
-            "exec line should use _NET_PREFIX array: {content}"
-        );
-        assert!(
-            content.contains(r#"exec "${_NET_PREFIX[@]}" umu-run"#),
-            "umu exec branch should also use _NET_PREFIX array: {content}"
-        );
-    }
-
-    #[test]
-    fn network_isolation_disabled_no_unshare_in_exec() {
-        let mut request = make_gamescope_request(
-            GamescopeConfig::default(),
-            TrainerLoadingMode::SourceDirectory,
-        );
-        request.network_isolation = false;
-        let content = build_trainer_script_content(&request, "Test Game");
-        assert!(
-            content.contains("# Network isolation: disabled"),
-            "script header should indicate disabled: {content}"
-        );
-        assert!(
-            !content.contains("unshare"),
-            "exec line should NOT contain unshare: {content}"
-        );
-    }
-
-    #[test]
-    fn network_isolation_with_gamescope_uses_both_prefixes() {
-        let mut request = make_gamescope_request(
-            GamescopeConfig {
-                enabled: true,
-                fullscreen: true,
-                ..Default::default()
-            },
-            TrainerLoadingMode::CopyToPrefix,
-        );
-        request.network_isolation = true;
-        let content = build_trainer_script_content(&request, "Test Game");
-        assert!(
-            content.contains(r#"exec "${_GS_PREFIX[@]}" "${_NET_PREFIX[@]}" "$PROTON" run"#),
-            "gamescope should be outermost, net isolation inside: {content}"
-        );
-        assert!(
-            content.contains(r#"exec "${_GS_PREFIX[@]}" "${_NET_PREFIX[@]}" umu-run"#),
-            "umu exec branch should use both prefix arrays: {content}"
-        );
-    }
-
-    #[test]
-    fn build_exec_line_emits_umu_probe_and_dual_exec() {
-        let request = make_gamescope_request(
-            GamescopeConfig::default(),
-            TrainerLoadingMode::SourceDirectory,
-        );
-        let content = build_trainer_script_content(&request, "Test Game");
-        assert!(
-            content.contains("command -v umu-run"),
-            "script should probe for umu-run: {content}"
-        );
-        assert!(
-            content.contains("_UMU_AVAILABLE=1"),
-            "script should mark umu availability: {content}"
-        );
-        assert!(
-            content.contains(r#"exec "$PROTON" run"#),
-            "fallback exec must remain: {content}"
-        );
-        assert!(
-            content.contains("exec umu-run"),
-            "umu exec branch must be emitted: {content}"
-        );
-    }
-
-    #[test]
-    fn proton_preference_bypasses_umu_run_for_exports() {
-        let mut request = make_gamescope_request(
-            GamescopeConfig::default(),
-            TrainerLoadingMode::SourceDirectory,
-        );
-        request.umu_preference = UmuPreference::Proton;
-        request.umu_game_id = "custom-77".to_string();
-        request.runtime_steam_app_id = "9000".to_string();
-
-        let content = build_trainer_script_content(&request, "Test Game");
-
-        assert!(
-            !content.contains("command -v umu-run"),
-            "Proton opt-out should not probe for umu-run: {content}"
-        );
-        assert!(
-            !content.contains("exec umu-run"),
-            "Proton opt-out should not emit umu-run exec: {content}"
-        );
-        assert!(
-            !content.contains("export GAMEID="),
-            "Proton opt-out should not export umu env: {content}"
-        );
-        assert!(
-            content.contains("unset GAMEID PROTON_VERB PROTONPATH"),
-            "Direct Proton export should clear umu env: {content}"
-        );
-        assert!(
-            content.contains(r#"exec "$PROTON" run "$trainer_host_path""#),
-            "Proton opt-out should preserve direct Proton exec: {content}"
-        );
-    }
-
-    #[test]
-    fn umu_export_emits_trainer_env_contract_with_runtime_precedence() {
-        let mut request = make_gamescope_request(
-            GamescopeConfig::default(),
-            TrainerLoadingMode::SourceDirectory,
-        );
-        request.umu_preference = UmuPreference::Umu;
-        request.steam_app_id = "12345".to_string();
-        request.runtime_steam_app_id = "67890".to_string();
-        request.umu_game_id = "custom-override".to_string();
-
-        let content = build_trainer_script_content(&request, "Test Game");
-
-        assert!(
-            content.contains("export GAMEID='custom-override'"),
-            "umu_game_id should outrank Steam and runtime ids: {content}"
-        );
-        assert!(
-            content.contains("export PROTON_VERB='runinprefix'"),
-            "exported trainer scripts must preserve trainer verb parity: {content}"
-        );
-        assert!(
-            content.contains("export PROTONPATH='/opt/proton'"),
-            "umu branch must export the Proton directory: {content}"
-        );
-        assert!(
-            content.contains("WARNING: umu preference requested but umu-run is unavailable"),
-            "explicit umu preference should warn before fallback: {content}"
-        );
-        assert!(
-            content.contains("unset GAMEID PROTON_VERB PROTONPATH"),
-            "fallback direct Proton path should clear umu env: {content}"
-        );
-    }
-}
