+++ src/crosshook-native/crates/crosshook-core/src/export/launcher/content.rs
@@ -1,0 +1,322 @@
+use crate::launch::runtime_helpers::build_gamescope_args;
+use crate::launch::script_runner::resolve_launch_proton_path;
+use crate::profile::GamescopeConfig;
+use crate::settings::UmuPreference;
+
+use super::paths::{desktop_icon_value, escape_desktop_exec_argument, shell_single_quoted};
+use super::types::SteamExternalLauncherExportRequest;
+
+pub(crate) fn build_trainer_script_content(
+    request: &SteamExternalLauncherExportRequest,
+    display_name: &str,
+) -> String {
+    let resolved_proton_path =
+        resolve_launch_proton_path(&request.proton_path, &request.steam_client_install_path);
+    let mut content = String::new();
+    content.push_str("#!/usr/bin/env bash\n");
+    content.push_str("set -euo pipefail\n\n");
+    content.push_str(&format!("# {display_name} - Trainer launcher\n"));
+    content.push_str("# Generated by CrossHook\n");
+    content.push_str("# https://github.com/yandy-r/crosshook\n");
+    content.push_str("# Launch this after the game has reached the in-game menu.\n");
+    content.push_str(&format!(
+        "# Trainer loading mode: {}\n",
+        request.trainer_loading_mode.as_str()
+    ));
+    content.push_str(&format!(
+        "# Network isolation: {}\n",
+        if request.network_isolation {
+            "enabled (unshare --net)"
+        } else {
+            "disabled"
+        }
+    ));
+    content.push_str(&format!(
+        "PREFIX_ROOT={}\n",
+        shell_single_quoted(&request.prefix_path)
+    ));
+    content.push_str(
+        r#"if [[ "${PREFIX_ROOT##*/}" == "pfx" ]]; then
+  export WINEPREFIX="$PREFIX_ROOT"
+  export STEAM_COMPAT_DATA_PATH="$(dirname "$PREFIX_ROOT")"
+elif [[ -d "$PREFIX_ROOT/pfx" ]]; then
+  export WINEPREFIX="$PREFIX_ROOT/pfx"
+  export STEAM_COMPAT_DATA_PATH="$PREFIX_ROOT"
+else
+  export WINEPREFIX="$PREFIX_ROOT"
+  export STEAM_COMPAT_DATA_PATH="$PREFIX_ROOT"
+fi
+"#,
+    );
+    if !request.steam_client_install_path.trim().is_empty() {
+        content.push_str(&format!(
+            "export STEAM_COMPAT_CLIENT_INSTALL_PATH={}\n",
+            shell_single_quoted(&request.steam_client_install_path)
+        ));
+    }
+    content.push_str(&format!(
+        "PROTON={}\n",
+        shell_single_quoted(&resolved_proton_path)
+    ));
+    content.push_str(&format!(
+        "TRAINER_HOST_PATH={}\n",
+        shell_single_quoted(&request.trainer_path)
+    ));
+    if request.gamescope.enabled {
+        content.push_str(&build_gamescope_script_block(&request.gamescope));
+    }
+    match request.trainer_loading_mode {
+        crate::profile::TrainerLoadingMode::SourceDirectory => {
+            content.push_str(
+                r#"
+
+trainer_host_path="$(realpath "$TRAINER_HOST_PATH")"
+trainer_host_dir="$(dirname "$trainer_host_path")"
+cd "$trainer_host_dir"
+
+"#,
+            );
+            content.push_str(&build_exec_line(request, "\"$trainer_host_path\""));
+        }
+        crate::profile::TrainerLoadingMode::CopyToPrefix => {
+            content.push_str(
+                r#"
+
+copy_support_directory_if_present() {
+  local source_dir="$1"
+  local target_dir="$2"
+  local child_name="$3"
+
+  if [[ -d "$source_dir/$child_name" ]]; then
+    mkdir -p "$target_dir"
+    cp -R "$source_dir/$child_name" "$target_dir/"
+  fi
+}
+
+stage_trainer_support_files() {
+  local trainer_source_dir="$1"
+  local staged_target_dir="$2"
+  local trainer_file_name="$3"
+  local trainer_base_name="$4"
+  local sibling_file
+  local sibling_name
+
+  shopt -s nullglob
+
+  for sibling_file in "$trainer_source_dir"/*; do
+    sibling_name="$(basename "$sibling_file")"
+
+    if [[ "$sibling_name" == "$trainer_file_name" ]]; then
+      continue
+    fi
+
+    if [[ -f "$sibling_file" ]]; then
+      case "$sibling_name" in
+        "$trainer_base_name".*.json|\
+        "$trainer_base_name".*.config|\
+        "$trainer_base_name".*.ini|\
+        "$trainer_base_name".*.dll|\
+        "$trainer_base_name".*.bin|\
+        "$trainer_base_name".*.dat|\
+        "$trainer_base_name".*.pak)
+          cp -f "$sibling_file" "$staged_target_dir/"
+          ;;
+        *.dll|*.json|*.config|*.ini|*.pak|*.dat|*.bin)
+          cp -f "$sibling_file" "$staged_target_dir/"
+          ;;
+      esac
+    fi
+  done
+
+  shopt -u nullglob
+
+  for support_dir in assets data lib bin runtimes plugins locales cef resources; do
+    copy_support_directory_if_present "$trainer_source_dir" "$staged_target_dir" "$support_dir"
+  done
+}
+
+trainer_host_path="$(realpath "$TRAINER_HOST_PATH")"
+trainer_file_name="$(basename "$trainer_host_path")"
+trainer_base_name="${trainer_file_name%.*}"
+trainer_source_dir="$(dirname "$trainer_host_path")"
+staged_trainer_root="$WINEPREFIX/drive_c/CrossHook/StagedTrainers"
+staged_trainer_dir="$staged_trainer_root/$trainer_base_name"
+staged_trainer_host_path="$staged_trainer_dir/$trainer_file_name"
+staged_trainer_windows_path="C:\\CrossHook\\StagedTrainers\\$trainer_base_name\\$trainer_file_name"
+
+mkdir -p "$staged_trainer_dir"
+cp -f "$trainer_host_path" "$staged_trainer_host_path"
+stage_trainer_support_files "$trainer_source_dir" "$staged_trainer_dir" "$trainer_file_name" "$trainer_base_name"
+
+"#,
+            );
+            content.push_str(&build_exec_line(
+                request,
+                "\"$staged_trainer_windows_path\"",
+            ));
+        }
+    }
+    content
+}
+
+fn build_gamescope_script_block(config: &GamescopeConfig) -> String {
+    let raw_args = build_gamescope_args(config);
+    let has_fullscreen = raw_args.iter().any(|arg| arg == "-f");
+
+    let mut quoted_args: Vec<String> = Vec::new();
+    if !has_fullscreen {
+        quoted_args.push(shell_single_quoted("-f"));
+    }
+    for token in &raw_args {
+        quoted_args.push(shell_single_quoted(token));
+    }
+    let args_literal = quoted_args.join(" ");
+
+    let mut block = String::new();
+    block.push_str(&format!(
+        "\n# Gamescope wrapper — launches trainer in its own compositor session.\n\
+         _GAMESCOPE_ARGS=({args_literal})\n"
+    ));
+
+    if config.allow_nested {
+        block.push_str("_GS_PREFIX=(gamescope \"${_GAMESCOPE_ARGS[@]}\" --)\n");
+    } else {
+        block.push_str(
+            r#"if [[ -z "${GAMESCOPE_WAYLAND_DISPLAY:-}" ]]; then
+  _GS_PREFIX=(gamescope "${_GAMESCOPE_ARGS[@]}" --)
+else
+  _GS_PREFIX=()
+fi
+"#,
+        );
+    }
+
+    block
+}
+
+fn build_exec_line(request: &SteamExternalLauncherExportRequest, target_path: &str) -> String {
+    let mut block = String::new();
+
+    if request.network_isolation {
+        block.push_str(
+            r#"_NET_PREFIX=()
+if unshare --net true >/dev/null 2>&1; then
+  _NET_PREFIX=(unshare --net)
+else
+  echo "[CrossHook] WARNING: unshare --net unavailable — launching without network isolation" >&2
+fi
+"#,
+        );
+    }
+
+    let prefix = match (request.gamescope.enabled, request.network_isolation) {
+        (true, true) => "\"${_GS_PREFIX[@]}\" \"${_NET_PREFIX[@]}\" ",
+        (true, false) => "\"${_GS_PREFIX[@]}\" ",
+        (false, true) => "\"${_NET_PREFIX[@]}\" ",
+        (false, false) => "",
+    };
+    let direct_proton_exec = format!(
+        "unset GAMEID PROTON_VERB PROTONPATH\nexec {prefix}\"$PROTON\" run {target_path}\n"
+    );
+
+    match request.umu_preference {
+        UmuPreference::Proton => {
+            block.push_str(&direct_proton_exec);
+        }
+        UmuPreference::Auto | UmuPreference::Umu => {
+            block.push_str(
+                r#"_UMU_AVAILABLE=0
+if command -v umu-run >/dev/null 2>&1; then
+  _UMU_AVAILABLE=1
+fi
+"#,
+            );
+            let umu_exports = build_umu_env_exports(request);
+            let fallback_warning = if request.umu_preference == UmuPreference::Umu {
+                "  echo \"[CrossHook] WARNING: umu preference requested but umu-run is unavailable — falling back to direct Proton\" >&2\n"
+            } else {
+                ""
+            };
+            block.push_str(&format!(
+                "if [ \"$_UMU_AVAILABLE\" = \"1\" ]; then\n{umu_exports}  exec {prefix}umu-run {target_path}\nelse\n{fallback_warning}  {direct_proton_exec}fi\n"
+            ));
+        }
+    }
+
+    block
+}
+
+fn build_umu_env_exports(request: &SteamExternalLauncherExportRequest) -> String {
+    let resolved_proton_path =
+        resolve_launch_proton_path(&request.proton_path, &request.steam_client_install_path);
+    format!(
+        "  export GAMEID={}\n  export PROTON_VERB='runinprefix'\n  export PROTONPATH={}\n",
+        shell_single_quoted(&resolve_umu_game_id_for_env(request)),
+        shell_single_quoted(&proton_path_dirname(&resolved_proton_path))
+    )
+}
+
+fn resolve_steam_app_id_for_umu(request: &SteamExternalLauncherExportRequest) -> &str {
+    let umu_override = request.umu_game_id.trim();
+    if !umu_override.is_empty() {
+        return umu_override;
+    }
+
+    let steam_id = request.steam_app_id.trim();
+    if !steam_id.is_empty() {
+        return steam_id;
+    }
+
+    let runtime_id = request.runtime_steam_app_id.trim();
+    if !runtime_id.is_empty() {
+        return runtime_id;
+    }
+
+    ""
+}
+
+fn resolve_umu_game_id_for_env(request: &SteamExternalLauncherExportRequest) -> String {
+    let trimmed = resolve_steam_app_id_for_umu(request).trim();
+    if trimmed.is_empty() {
+        "umu-0".to_string()
+    } else {
+        trimmed.to_string()
+    }
+}
+
+fn proton_path_dirname(proton_path: &str) -> String {
+    std::path::Path::new(proton_path.trim())
+        .parent()
+        .map(|parent| parent.to_string_lossy().into_owned())
+        .unwrap_or_default()
+}
+
+pub(crate) fn build_desktop_entry_content(
+    display_name: &str,
+    slug: &str,
+    script_path: &str,
+    launcher_icon_path: &str,
+) -> String {
+    let mut content = String::new();
+    content.push_str("[Desktop Entry]\n");
+    content.push_str("Type=Application\n");
+    content.push_str("Version=1.0\n");
+    content.push_str(&format!("Name={display_name} - Trainer\n"));
+    content.push_str(&format!(
+        "Comment=Trainer launcher for {display_name}. Generated by CrossHook: https://github.com/yandy-r/crosshook\n"
+    ));
+    content.push_str(&format!(
+        "Exec=/bin/bash {}\n",
+        escape_desktop_exec_argument(script_path)
+    ));
+    content.push_str("Terminal=false\n");
+    content.push_str("Categories=Game;\n");
+    content.push_str(&format!(
+        "Icon={}\n",
+        desktop_icon_value(launcher_icon_path)
+    ));
+    content.push_str("StartupNotify=false\n");
+    content.push_str(&format!("X-CrossHook-Profile={display_name}\n"));
+    content.push_str(&format!("X-CrossHook-Slug={slug}\n"));
+    content
+}
